Ignore ERROR_SUCCESS in Windows InitCommonControlsEx()
As per typical Microsoft behavior, `InitCommonControlsEx` appears to sometimes fail with an error of ERROR_SUCCESS. The library totally works fine in this case. I have not investigated really deeply but it seems this fixes init in this case. Platform is Windows 10 x64.

It's being used from a Rust app so it's possible that this is happening because Rust's `winapi` crate is calling another initializer or something, but I don't know. Just know that the library works fine if you ignore an error of "SUCCESS."