darwin: Fix uiCombobox append/insert/delete/clear/selected.
Fix the default Selected() to return `-1` like other
platforms.

Fix the behavior of uiCombobox for Append(), Insert(),
Delete(), and Clear() so that the Selected() item index
gets updated as needed.

Additionally disable the control when there are 0 items
in the uiCombobox matching unix behavior. Otherwise the
user is able to select an empty field which is confusing.