doc: Add a comment that `window` may be `NULL`.
Apparently uiMenuItemOnClicked() may pass `NULL` as the `window`
parameter under certain circumstances on darwin. At least that
possibility is suggested here:
- darwin/window.m:461
- darwin/window.m:142

I added a @todo to the docs as well, as I am unaware of how to
trigger such behavior. The code does handle that possibility though,
so users need to be aware of the possibility to not trigger any
`NULL` pointer derefs.